Framework for fixing overflow issues
When the number of decimal places `f` produces a coefficient `10^f`
which is not representable in the container type `T` we can run into
overflow issues. Additionally, if the FixedDecimal value is also large
we can also run into overflow when working with a representable
coefficient.

The new function `coefficient` provides us with a representable power of
10 which uses a type which will never overflow when multiplied by the
value.