Speed up * for FD{Int64} w/ custom fldmod_by_const
Add a custom implementation of `fldmod(x::T,y)` for when `y` is a const,
and call it whenever T >= Int128.

This improves performance for `*(a::FD{Int64}, b::FD{Int64})`, b/c `*` passes
the result of `widemul(a,b)` into `fldmod` -- and LLVM isn't able to
optimize `fldmodinline(::Int128,::Const{Int128})` as well as we'd like.

Here, we opt to use a custom implementation of `fldmod` in that case,
which runs significantly faster, making the multiply faster.

--------

In this commit, Int128 still widens to BigInt, and we do not use the
custom implementation in that case.

In a follow-up commit, we'll change Int128 to widen to a simple 256-bit
integer implementation, so that we can use the custom fldmod.