Finish removing the BigInts from * for FD{Int128}!
Finally implements the fast-multiplication optimization from
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/45, but this
time for 128-bit FixedDecimals! :)

This is a follow-up to
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/93, which
introduces an Int256 type for widemul. However, the fldmod still
required 2 BigInt allocations.

Now, this PR uses a custom implementation of the LLVM div-by-const
optimization for (U)Int256, which briefly widens to Int512 (😅) to
perform the fldmod by the constant 10^f coefficient.

This brings 128-bit FD multiply to the same performance as 64-bit. :)